fixing tests for sm100
