+++ tests/testutils.py
@@ -39,6 +39,22 @@ from packaging import version
 import torch_harmonics.distributed as thd
 
 
+def _is_sm90():
+    """Return True when the default CUDA device is Hopper (SM 9.0)."""
+    if not torch.cuda.is_available():
+        return False
+    major, _ = torch.cuda.get_device_capability()
+    return major == 9
+
+
+def _is_sm100():
+    """Return True when the default CUDA device is Blackwell (SM 10.0)."""
+    if not torch.cuda.is_available():
+        return False
+    major, _ = torch.cuda.get_device_capability()
+    return major == 10
+
+
 def set_seed(seed=333):
     """Set the torch + CUDA random seed.
